build: update gradle wrapper scripts
The "gradlew" and "gradlew.bat" scripts have been updated to improve the startup process and handle maximum file descriptor limits. This includes changes to the way JVM options are added to the script, as well as checks for the availability of the "xargs" command. These updates should enhance the script's reliability and performance.