ci: update build GitHub Actions Workflow for testing and preparing plugin release
This commit adds a GitHub Actions Workflow in the `.github/workflows/build.yml` file. The workflow is created for testing and preparing the plugin release in several steps:

- Validate Gradle Wrapper
- Run 'test' and 'verifyPlugin' tasks
- Run Qodana inspections
- Run 'buildPlugin' task and prepare artifact for further tests
- Run 'runPluginVerifier' task
- Create a draft release

The workflow is triggered on `push` and `pull_request` events. For more information, refer to the GitHub Actions reference: [link](https://help.github.com/en/actions).

The commit also includes various changes to setup the environment, export properties, clean up unnecessary files, fetch sources, and cache the Plugin Verifier IDEs. Additionally, it builds the plugin, prepares the plugin artifact, and uploads it as an artifact for downloading. Finally, the commit adds steps to run the plugin structure verification along with the IntelliJ Plugin Verifier and Qodana code inspections.

These changes are made to automate the process of testing and preparing the plugin release in a standardized and controlled manner.