chore: simplify SVG code and remove unused Icon
Simplified the code for SVG files (type.svg, gatsby.svg, pluginIcon.svg) by reducing unnecessary metadata and consolidating the paths for the shapes. The 'FileType' field in GatsbyIcons.kt was also removed because it was fetching the 'type.svg' icon which is no longer in use. This cleanup process helps improve the performance and readability of the code.