refactor: update project generator for code style consistency
A code refactor was performed on the GatsbyCliProjectGenerator class for the sake of code style consistency. Mainly, the usage of ":" in function declarations has been omitted when it can be inferred by the compiler. Moreover, instead of directly importing and calling GatsbyBundle.message(), a top-level import operation is done for GatsbyBundle.message and it is called as a function for better readability. These changes aim to improve code readability without modifying any functionality or logic.